Revert "fix: revert openapi version"
This reverts commit 44c54e29ad62916df31752e949cbab344c0afb38.

+++ core/src/main/kotlin/com/rarible/flow/core/converter/OrderToDtoConverter.kt
@@ -44,7 +44,7 @@ class OrderToDtoConverter(
             }
 
             return FlowOrderDto(
-                id = source.id.toLong(),
+                id = source.id,
                 itemId = "${source.itemId}",
                 maker = source.maker.formatted,
                 taker = null,
